pyproject.toml
[tool.poetry]
name = "quant-platform"
version = "0.1.0"
description = "Modular LLM-native Quant Research & Trading Platform"
authors = ["Rodrigo <rodrigosmgm@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.1.1"
python-json-logger = "^3.3.0"
alembic = "^1.16.4"
psycopg2-binary = "^2.9.10"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
email-validator = "^2.2.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pandas = "^2.3.1"
httpx = "^0.28.1"
yfinance = "^0.2.65"
sqlalchemy = "^2.0.41"
redis = "^6.2.0"
openai = "^1.97.1"
anthropic = "^0.60.0"
tiktoken = "^0.9.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"
mypy = "^1.17.0"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
